Doc added for user API aand ascent API. ENV variables now used. Feedback from Pull Request implemented

+++ typescript implementation/database-service/src/database/databaseOperation.ts
@@ -1,49 +1,51 @@
-import { Mongodb } from './mongodb'
-import { DeleteWriteOpResultObject} from "mongodb"
+import { Mongodb } from "./mongodb";
+import { DeleteWriteOpResultObject } from "mongodb";
 
-interface databaseQuery { 
-    [key: string]: number | string
+interface databaseQuery {
+  [key: string]: number | string;
 }
 
-export async function searchInSummit(search: string, value: number | string, collection: string, database: string, useQuery: boolean): Promise<Array<any>> {
-    
-    var query: databaseQuery = {};
+export async function searchInSummit(
+  search: string,
+  value: number | string,
+  collection: string,
+  database: string,
+  useQuery: boolean
+): Promise<Array<any>> {
+  var query: databaseQuery = {};
 
-    if (useQuery)
-        query[search] = value;
+  if (useQuery) query[search] = value;
 
-    const cursor = await Mongodb.client.db(database).collection(collection).find(query);
+  const cursor = Mongodb.client.db(database).collection(collection).find(query);
 
-    let result: any = [];
+  let result: any = [];
 
-    return new Promise<Array<any>>(async (resolve, rejects) => {
-
-        await cursor.forEach((entry) => {
-
-            result.push(entry);
-        });
-        resolve(result);
+  return new Promise<Array<any>>(async (resolve, rejects) => {
+    await cursor.forEach((entry) => {
+      result.push(entry);
     });
+    resolve(result);
+  });
 }
 
-export async function deleteInSummit(search: string, value: number | string, collection: string, deleteAll: boolean): Promise<number> {
-    
-    var query: databaseQuery = {};
+export async function deleteInSummit(
+  search: string,
+  value: number | string,
+  collection: string,
+  deleteAll: boolean
+): Promise<number> {
+  var query: databaseQuery = {};
 
-    query[search] = value;
-    
-    let collectionDB = Mongodb.client.db("summit").collection(collection);
+  query[search] = value;
 
-    let cursor: DeleteWriteOpResultObject;
+  let collectionDB = Mongodb.client.db("summit").collection(collection);
 
-    if(deleteAll)
-        cursor = await collectionDB.deleteMany(query);
-    else
-        cursor = await collectionDB.deleteOne(query);
+  let cursor: DeleteWriteOpResultObject;
 
+  if (deleteAll) cursor = await collectionDB.deleteMany(query);
+  else cursor = await collectionDB.deleteOne(query);
 
-    return new Promise<number>((resolve, reject) => {
-
-        resolve(cursor.deletedCount);
-    });
+  return new Promise<number>((resolve, reject) => {
+    resolve(cursor.deletedCount);
+  });
 }
